rename component for clarity
CameraSettingsView is now CameraReviewSettingsView
